Fixing some warnings
https://github.com/bibledit/cloud/issues/788

+++ bb/css.cpp
@@ -92,7 +92,7 @@ string bible_css (void * webserver_request)
     float letterspacing = convert_to_float (request->post["letterspacing"]);
     if (letterspacing < -3) letterspacing = -3;
     if (letterspacing > 3) letterspacing = 3;
-    Database_Config_Bible::setLetterSpacing (bible, 10 * letterspacing);
+    Database_Config_Bible::setLetterSpacing (bible, static_cast<int>(10 * letterspacing));
     
     page += Assets_Page::success ("The information was saved.");
